+++ CHANGELOG.md
@@ -49,7 +49,7 @@ Properties are mostly reflected 1:1 from OpenLayers as Component Properties for 
 - 👩🏻‍🔧 migrated all components to use [`<script setup>` syntax](https://vuejs.org/api/sfc-script-setup.html)
 - 🗺️ aligned API with [OpenLayers 7.4.0](https://github.com/openlayers/openlayers/releases/tag/v7.4.0)
 - 📘 updated & restructured the whole documentation
-- 🧩 added new components: `ol-layer-group`, `ol-source-tile-json`, `ol-source-stamen` and `ol-vector-image-layer`
+- 🧩 added new components: `ol-layer-group`, `ol-source-tile-json`, `ol-source-stadia-maps` and `ol-vector-image-layer`
 - 🚀 expose all interactions to be able to call access them via `ref` in templates
 - 🐞 lot's of bug fixes
 - 📖 improved the documentation